+++ TaskMenu/Models/AppState.swift
@@ -43,14 +43,6 @@ final class AppState {
             }
         }
     }
-    var experimentalFullWindowLiquidGlassEnabled: Bool {
-        didSet {
-            userDefaults.set(
-                experimentalFullWindowLiquidGlassEnabled,
-                forKey: Constants.UserDefaults.experimentalFullWindowLiquidGlassEnabledKey
-            )
-        }
-    }
 
     var selectedList: TaskList? {
         taskLists.first { $0.id == selectedListId }
@@ -236,9 +228,6 @@ final class AppState {
         self.dueDateNotificationsEnabled = userDefaults.object(
             forKey: Constants.UserDefaults.dueDateNotificationsEnabledKey
         ) as? Bool ?? true
-        self.experimentalFullWindowLiquidGlassEnabled = userDefaults.object(
-            forKey: Constants.UserDefaults.experimentalFullWindowLiquidGlassEnabledKey
-        ) as? Bool ?? false
         self.isSignedIn = authService.isSignedIn
     }