+++ TaskMenu/Utilities/Constants.swift
@@ -37,7 +37,6 @@ enum Constants {
 
     enum UserDefaults {
         static let dueDateNotificationsEnabledKey = "dueDateNotificationsEnabled"
-        static let experimentalFullWindowLiquidGlassEnabledKey = "experimentalFullWindowLiquidGlassEnabled"
     }
 
     enum Notifications {